MINOR CHANGE: Fine Tuning Model Parameters
Started fine tuning parameters to make the text smoother in terms of
sentence structer. Managed to develop a cohesive model that is able to
develop actual sentences. The only issue is the problem with relevance
in relation to the previous text. Most stories created appear to have
random sentences placed together.