CHANGE: Parameters and New Iteration
Created a global variable within the pipeline library to determine the
version of the model. Created a for loop to test the text generation at
different temperatures. From the observations made, it seems that the
ideal temp is 0.8. This may be attributed to the parameters of the model
still requiring further fine-tuning, which may be improved upon
throughout the later versions (Currently on the 20th version of the
model).

Improved upon the model architecture by limiting the number of LSTM
layers down to two. Multiplied the number of units for the second LSTM
layer by a factor of 4. Limited the number of epochs to 130, this is
mainly attributed to the fast learning rate of the adam optimizer.