CHANGE: ChatGPT Assisted Attempt
Attempted to follow through some code written by chatgpt in conjunction
with the tutorial set in the tensorflow page. This failed
catastrophically as the current algorithm is not able to develop an
appropriate environment for training based on the observation spec
changing after every word. Will have to develop multiple environments
upon which one could train the model.